CustomButtonConfig initialization
- moved getServerConfig() dependency to a new file to fix dependency issue with tests
- removing unnecessary imports
